requirements.txt
asgiref==3.8.1
certifi==2024.12.14
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
cloudinary==1.42.1
dj-database-url==2.3.0
Django==5.1.5
django-cloudinary-storage==0.3.0
djangorestframework==3.15.2
idna==3.10
msgpack==1.1.0
psycopg2-binary==2.9.10
python-decouple==3.8
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
